backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.1
networkx==3.4.2
torch==2.3.1
torchvision==0.18.1
trimesh==4.4.7
pymeshlab==2023.12
huggingface_hub==0.25.1
numpy==1.26.4
Pillow==10.4.0
scikit-image==0.24.0
opencv-python==4.10.0.84
einops==0.8.0
omegaconf==2.3.0
jaxtyping==0.2.28
typeguard==4.3.0
peft==0.11.1
diffusers==0.30.2
transformers==4.43.4
accelerate==0.33.0
rembg==2.0.56
uvicorn
